Adding a field to 'event' to record which organization organized a workshop.
We need to track which organization did the admin for a workshop (e.g., SSI organizing in the UK, MSL organizing in North America) so that fees can be divided up.  We do this by having another foreign key from the `event` table back to the `site` table, and adding the partner organizations as sites.