workshops/views.py: Add person.num_taught per person
My previous Count() suggestion implemented in 03ecafeb (Annotating
instructors with number of times taught, 2014-12-18) was removing
folks who had never taught.  This commit replaces that approach with
separate database calls for each person, which is less efficient but
works for folks who have never taught.  An alternative approach would
be to use a custom filter [1] to add this information at the UI level.
For an implementation of that approach see [2].  We've gone with the
in-view calculation here because Greg expects we'll be filtering by
this attribute in the future [3].

[1]: https://docs.djangoproject.com/en/1.7/howto/custom-template-tags/#writing-custom-template-filters
[2]: https://github.com/gvwilson/amy/pull/7
[3]: https://github.com/gvwilson/amy/pull/7#issuecomment-67651211