workshops.views: Edit awards in the person-detail edit view.
Using django-extra-views to automatically deploy the inline formsets
[1].  To handle the datepicker stuff, you need my widgets addition
[2].  I hope that lands soon (it's a simple change), but there are a
number of older issues and pull-requests in that repository, so I'm
not sure how active the upstream maintainer is.  If we get tired of
waiting for a release, we can for the library and depend on our fork,
but I'd rather avoid that if possible.

The datepicker widgets are from bootstrap-datepicker via the data-api
instantiation [3].

This commit also tweaks the person-detail display to show a person's
current awards.

[1]: https://pypi.python.org/pypi/django-extra-views/
[2]: https://github.com/AndrewIngram/django-extra-views/pull/93
[3]: https://bootstrap-datepicker.readthedocs.org/en/release/#data-api