workshops.urls: Add 'POST /workshops/event/{slug}/award'
To automatically create awards based on an event.  It's useful for
badging organizers (and once we have a badge for them and a list of
actual attendees, for badging students).  I'm not sure about badging
instructors this way.  The current implementation does badge
instructors for the event that weren't already badged, but maybe we
want to reserve badging for folks completing the instructor training
course (regardless of whether or not they've been an instructor at an
event)?  In that case we'd want to tie the instructor award creation
to setting Person.trainee.status to the 'completed' status.  We may
also want to distinguish badging logic based on Event.project (SWC,
Data Carpentry, ...).

I've removed the HttpResponseRedirect import because we don't use it
since the last reference was removed in 4e211e12 (First working
version of instructor lookup form, 2014-12-16).