Fix: non-blank fields preventing from saving new Event
Fields marked as `null` can save this value to the database.  However,
when it comes to validation, we need to explicitely allow blank form
fields for NULL-able model fields:
https://docs.djangoproject.com/en/1.7/ref/models/fields/#null

One curious case was `notes` field. I couldn't get it to work without
`blank`, even though it has a `default=""`.
All in all, it worked.