workshops/test: Remove obsolete references to Project
It looks like the Project entries were just because Event.project
hadn't been nullable.  The new many-to-many Event.tags is nullable,
since there is no null constraint for ManyToManyField.  That means we
can create Events for testing without a dummy Tag, so that's what I've
changed the tests to do.