New view: bulk people csv upload confirmation page
CSV data no longer end up in a database right away, but are stored in
session.  The confirmation page shows that data fetched from the
session.  Additionally, a verification on the data happens.

Later it will be possible to edit the values on-the-fly, ie. in the
confirmation page itself.  Not yet, though.