Bulk-edit individual fields on confirmation page
Bulk upload confirmation page allows for editing individual table cells.
As a reaction to mouse double-click, an input is shown and the cell's
text disappears.

On the backend side, all fields are retrieved from POST data.  Then
the data is formed into it's previous shape (list of dictionaries) and
saved to the session.

Then one of 3 actions (depending on what button user clicked: confirm,
cancel or verify) happens:

* verify: the data is being checked for errors
* cancel: session is cleared, nothing stored to the DB
* confirm: data is being saved in an transaction; if errors occur,
  proper message is displayed.