Class-based UpdateView with Context Mixin
Because of using class-based view `UpdateView` to update various objects
in AMY, we wanted to show these objects string representation in the
`<title>` of the page.  However, it's not possible to include context
variables in "pure" `UpdateView`.  It's possible with help from
`ContextMixin`.

This change introduces `UpdateViewContext` that's a child of both
`UpdateView` and `ContextMixin` and automatically sets `title` context
variable.