workshops/urls.py: Use IDs in the task detail/edit URLs
The old:

  task/{event_slug}/{person_id}/{role_name}

URLs are almost human readable.  If they were:

  task/{event_slug}/{person_slug}/{role_name}

I'd stick with the long form.  However, once you put an ID in the URL,
I think it might as well be the task ID.  We're not comfortable
requiring person slugs at the moment [1,2,3,4], so I think the task ID
URLs are the cleanest choice.

[1]: https://github.com/swcarpentry/amy/pull/125#issuecomment-69049496
[2]: https://github.com/swcarpentry/amy/pull/125#issuecomment-69143688
[3]: https://github.com/swcarpentry/amy/pull/125#issuecomment-69208785
[4]: https://github.com/swcarpentry/amy/pull/125#issuecomment-70397304