workshops/util.py: Expect a Unicode stream in upload_person_task_csv
Instead of passing in an encoding and doing some issubclass stuff,
just require a Unicode stream in upload_person_task_csv.  In general,
decoding to Unicode as soon as possible and encoding to bytes as late
as possible makes for cleaner code (who better to know the proper
encoding than the first/last code that touches the content?).  We see
that here, with a simpler upload_person_task_csv implementation and
only a two more lines in person_bulk_add to wrap the incoming file in
an TextIOWrapper.

The accept-charset business is a hack to work around shortcomings in
modern browsers.  For example, the data posted from my Firefox 24.8
looks like:

  Content-Type: multipart/form-data; boundary=---------------------------196554893016147250581211961313
  Content-Length: 477

  -----------------------------196554893016147250581211961313

  Content-Disposition: form-data; name="csrfmiddlewaretoken"

  D8gooO2bYBXwsVQndknbnHUkJbs9ZRHp

  -----------------------------196554893016147250581211961313

  Content-Disposition: form-data; name="file"; filename="BulkPersonAddTemplate.csv"
  Content-Type: text/csv

  personal,middle,family,email,event,role
  αβ,γδ,Ε,a@b.com,2015-02-05-esu,student

  -----------------------------196554893016147250581211961313--

It's setting a per-file Content-Type (yay!), but that Content-Type
doesn't have charset information (boo!).  It looks like the Mozilla
folks tried to fix this [1], but backed it out in 2007 [2] because it
broke some crummy parsers [3,4].

I see nothing in RFC 2388 that is file-specific, but their only
Content-Type example does show a charset [5].  The only workable,
spec-compliant solution I can come up with is to explictly set
accept-charset [6].  Otherwise, the user agent is supposed to use
some wonky fallbacks [7] (although if DEFAULT_CHARSET was UTF-8,
those would end up as UTF-8 too).

The current approach is to respect the charset set by the user-agent
(which we can access via the .charset property [8]).  If it doesn't
set one, we expect it to respect the form's accept-charset and use
DEFAULT_CHARSET.

[1]: https://bugzilla.mozilla.org/show_bug.cgi?id=116346
[2]: https://bugzilla.mozilla.org/show_bug.cgi?id=116346#c79
[3]: https://bugzilla.mozilla.org/show_bug.cgi?id=116346#c74
[4]: https://bugzilla.mozilla.org/show_bug.cgi?id=392982#c45
[5]: https://tools.ietf.org/html/rfc2388#section-4.5
[6]: http://www.w3.org/TR/html5/forms.html#attr-form-accept-charset
[7]: http://www.w3.org/TR/html5/forms.html#multipart/form-data-encoding-algorithm
[8]: https://docs.djangoproject.com/en/1.7/ref/files/uploads/#django.core.files.uploadedfile.UploadedFile.charset