test_person.py: check for bool values in template
In person detail view we started displaying boolean field values as
either "yes" or "no".

Tests were covering only "True" or "False"; this is now fixed by
introducing a special conditional case for testing boolean field value:
if it's a boolean, then we check it's value for "yes" or "no", too.