Fix: add default value for empty cells in bulk upload
If we upload a record that does not have any value in specific cell, and
we want to change that value to something meaningful later on, we should
be able to click onto some text.  Default value (long dash) makes it
easy to click and clearly shows that the cell has no value.