test_util.py: upload existing user and assign existing task
This commit introduces a new test that makes sure uploading existing
user and assigning existing task to them fails.

Currently this test shows (through failing) that we allow multiple
the same tasks on the same event for the same user.  If this is done by
purpose, I'll revert this commit.