Fix: empty personal/family names representation
This fixes an error that caused hidden <input> elements to contain
`value="None"` instead of `value=""`.

Additionally, this commit changes the way we handle incoming form data.
In our models, `Person` has fields `personal`, `family` that are not
nullable.  Therefore we should not default to `None` values for these
fields, because it screws up `util.create_username`.