Fix: uploading malformed CSV with correct header
This resolves #231
(Bulk-upload: Not checking whether it is a valid csv file).

The issue occurred when user wanted to upload a file that had
a correct CSV header as a first row, while the rest of the file were not
comma separated values.

Such a file behaves like a CSV file with missing columns in the body.