Add a new form helper
Old bootstrap helper was turned into a class, and a new helper was
introduced (thanks to subclassing).
The new helper doesn't use <form> - this is because on event edit page
we have two separate forms that should be submitted simultaneously.

The new helper, BootstrapHelperWithoutForm, is only used on that event
edit page.