Add Badge.get_absolute_url, change Badge.__str__
`Badge.get_absolute_url` method was missing and proved to be useful in
cases when we want to add a link back to the badge details page.  Usage
comes in one of follow commits.

`Badge.__str__` on the other hand was returning only Badge names, not
their titles.  Title is more user-friendly than the name, right?