New view: person_edit (was PersonUpdate)
The former view was class-based and supported only singular form, while
we need two forms (one for editing person, one for adding new award to
that person).

This approach unfortunately breaks the tests, as they're expecting only
one `<form role="form">` in the source, but from the backend point of
view: it's easier and nicer to implement than one-form-approach
I decided to use on event edit page.