Shift task/award forms over the corresp. tables on event/person edit views
Both event-edit and person-edit pages display list of tasks/awards
associated with specific event/person.
For events, the list of learners tends to be very long, so I moved the
task form up.
Since person-edit page is almost like a mirror of event-edit page,
I did the same there.

The bootstrap helper was modified to use submit with value "Add" instead
of "Submit".