Add *_delete views
Allow airports, persons and sites to be hard-deleted, ie. removed from
the database (events and tasks are marked as deleted and not really
removed).

Buttons to remove specific person/airport/site appears on the bottom of
details page for that object.

A generic failed_to_delete.html template is used when deletetion breaks
our integrity constraints. It displays information that should
let the user decide what they need to delete first.