Add test for editing user lessons
This test makes sure we handle removal/addition of new qualifications
(M2M intermediate model between Person and Lesson).

There was an issue with TestBase._get_selected: it was returning strings
instead of lists, so it was difficult to reuse selected options from the
webpage.  I changed it to return lists.