'Recently changed' on dashboard view
This shows last 50 revisions on the dashboard.
Revisions are on for Event and Person models, without following any
related objects.

This means adding a badge to the person or changing an award will not
show in the 'Recently changed' list!  There was too much confusion with
changing related objects that I decided to drop it.