Show changes between revisions for some objects
This commit adds a view for showing changes between current revision
and latest known revision for some object.
We currently support revisions only for events and persons.
The changes are shown using Google's diffmatchpatch library.

As in previous commit: we won't show changes to related objects because
it was pain to figure out properly.
For example if some event was tagged another tag, it won't show up as
a change anywhere.