Move merging persons code to the workshops/util.py
The new function name is `merge_persons`.  This function deals with the
issues of updating only part of related objects.

There are unique constraints in Award and Task models, so we can only
update some awards/tasks and remove the rest.
The conflicts happen only when two persons have "the same" awards, ie.
when we're trying to assing the same badge to the same person twice.