Add 'Delete task' to the task detail page
Fixes #403.

Additionally to unify URLs for events, we default to event.get_ident
(which is either slug or ID) for URLs.

Because task deletion page should redirect back either to event edit
page or to all tasks page, I added a second URL (with 1 parameter more:
event_ident), for the same name ('task_delete').  So if the event_ident
is present we know we want to be redirected back to the event edit page.