Add tabs to the person-edit page
The page had became too crowded (3 forms, 2 listings).

I decided to give it a tabbed panes.
Tabs: Person, Awards, Tasks.
The first one allows for changing person, the second one shows person's
awards and allows user to add more, the third one shows person's tasks
and allows user to add more.

There is a downside to this solution: after submitting an incorrect
award or task form, the user is redirected to the person-edit page (ie.
person form shows up, and not the form with errors).

Also, the tabbed panes from Bootstrap use anchored links, but don't
switch to the selected pane if it's anchor is added to the requested
URL.