Add 'affiliation' to some templates
Additionally, some fields were moved. Now we should have the same order
of fields on all pages and forms.

This commit also adds 'affiliation' field to the Find Instructors page,
because the feature was requested by Tracy Teal, and I personally hope
to make that page more useful for her.