Make Event's admin_fee and attendance non-negative
This fixes #435.

Fixing Event.attendance was easy since we have a PositiveIntegerField
in Django already.

We cannot have a PositiveDecimalField, though.  I used a trick, ie.
Django validator for checking if the value is at least 0.  This trick
only works with forms, so if someone adds a negative admin_fee from the
CLI, then it will work.