Better filter by countries for All Hosts page
This filter shows country names only for countries used in the
Host.country field.

It looks like a hack, but it's kind of necessary. We're using
django-countries mechanism for limiting possible countries.
Countries are stored in the DB as 2-character strings (e.g. 'PL' for
Poland, 'GB' for United Kingdom).