Gray out, not hide links to actions that require permissions
This is done via if-else-endif for now, but the change wasn't so bad.
To save 5 lines of code, we might add some template tags that would
display the buttons in disabled state if required, but this refactoring
wasn't all that bad, so the tags weren't necessary.

The only thing impossible to gray-out was glyphicon links (pencils) to
object-edit views. They remain hidden for now.