Adding report of missing locations to problems page.
This turns the problems report into a multi-column display.  Unfortunately, the bodies of the email messages are now all on one line because I can't figure out how to get newlines into mailto href's in template strings.  (The old version passed them via context and relied on urlencode, which meant there was text in views.py that belonged in the template.)