Making form purpose clearer in titles.
The forms for requesting SWC and DC workshops didn't say which kind of workshop it was - this adds that.
It also make the title of the instructor info update form clearer.

Question: why are the SWC and DC workshop request forms classes, while the instructor info update form is a function?