all activity report: change instructor/learner counting
Now we count instructors per SWC and per DC workshops, and learners per
SWC and per DC workshops.

Learners figures are essentially sums of attendances.

Changing JS to properly display JSON output from API was a little tricky
because JS doesn't help us when we have to check if something is
JavaScript object… (for example it turned out that Arrays are Objects).