Fix test_number_of_attendees_increasing: speed it up
This test was using one redundant call to `event_edit` to get the page,
parse it, find forms, and use the second form values to our advantage.

This was completely unnecessary, as we can define form values ourselves.

The test went from 0.9s to about 0.2s (we probably can't do better
because we're testing view's functionality here).