Fix: 'ni-{member,historic}' to 'na-{member,historic}'
Switch event invoice status for "ni-{member,historic}" to "na-*"
("not invoiced" to "not applicable").

Add migration for rewriting old values in the database.

This migration also adds a missing migration for User model change
(username validator was extended to handle dashes too, so not a big
change in regards of migration).

This fixes #669.