Events merge: display event selection form
The form is served from the same address as the table with events
comparison and strategy selection. It's achieved by using GET
parameters.

So user selects two events and submits. The form is set to use GET, so
two slugs and two IDs end up in the address query.
Since they're available, the view won't display selection form but
proceed to finding events and then displaying their comparison.

Finally, if it has gotten a POST, then it tries to merge the events.