Duplicates: add link to people merging view
This isn't the best way to achieve the effect I wanted, but it's
the closest.

Partial issue is that sometimes there are event 3 duplicates in the
database - but most of the time it's only 2.

We only do 2-way merge so I decided to cheat a little and allow to merge
every person with the person above them. Since we have good ordering
(family, personal, email) then I feel safe…