Fix migration 0088*
Migration 0088* was responsible for making `Event.slug` a required
field.  This resulted in a data migration: some events that didn't have
any slug needed one.

In testing everything worked fine, but in production 2 events ended up
with the same slug "none-".  I decided to fix the migration in place by
adding a random suffix to the event's slug.