Refactor crispy form helpers
Changes in form helpers:

- Replace BootstrapHelperWithAdd with BootstrapHelper(submit_label='Add')

- Replace BootstrapHelperWiderLabels with BootstrapHelper(wider_labels=True)

- Replace BootstrapHelperGet with BootstrapHelper(use_get_method=True)

- Add BootstrapHelper.duplicate_buttons_on_top flag

  When this flag is set to True, submit buttons are displayed on both
  top and bottom of a form. The default value is False.

- Add other arguments to BootstrapHelper.__init__

Additionally, we now use form.helper instead of form_helper.

So far, FormHelper and Forms were created in a view independently, put
into a context and passed to a template where they were connected:

    {% crispy form form_helper %}

Now the helper is set directly on the forms. You don't have to think
about the helper in a view or in a template (unless you want it to be
request-dependent, i.e. when you want form_action to depend on the
request).

One more change: duplicated code (sidebar) in templates was extracted
and moved to base template.