Update DC self-organized workshop registration form
This fixes #967 3 issues:
* incorrect title
* wording in topics selection
* missing "Don't know yet" option in data analysis levels question

The two other issues, ie. missing "Other" option in M2M questions
(Domain for the workshop, Topics to be taught), won't be implemented
because the "Other domain/topic" field is placed right under mentioned
fields. It's not optimal setup, but we can't do much about it for now…