Add privacy consent field to multiple forms
This fixes #792 by adding a separate field (via a mixin) to multiple
forms that should ask the user for a privacy policy agreement.

This approach doesn't store any additional data in the DB, but forces
user to agree on the privacy policy instead. Such behavior may not be
desired (ie. do we by default assume everyone already in the database
agreed on that policy?).