Fix #981 -- add more ways to issue a badge
"SWC" and "DC" labels in "eligible" column in trainees list view are
now clickable when a person is eligible to become SWC/DC instructor, but
doesn't have a swc/dc-instructor badge yet. The user is moved to person
edit view with prefilled in "badge" and "event" fields. Here, the badge
can be awarded.