Change validation rules of TrainingRequest.state
Until now, pending requests can't be matched with a person and accepted
requests must be matched with a person.

Now, pending requests can't be matched with a *training* and accepted
requests may or may not be matched with a training. It doesn't matter
whether a request is matched with a person.