Bulk upload: find & use matching people
This commit introduces a new view intended for saving a selected match.

The changes in `verify_upload_person_task` are dictated by the shift in
workflow. Now we don't try to magically guess if user wants to re-use
an existing person. We don't match them by emails, either. The new
approach allows users to select a matching person, and for us,
developers, this means an easier and cheaper to maintain function.