verify_upload_person_task: fix verification error when reusing person
Chris Medrela noticed that when we're reusing an existing Person for
specific entry, and later user makes changes to that entry, in one part
of `verify_upload_person_task` the modified values are used, and in
other it uses DB values.  DB values should be used if "Use this person"
was selected.