Admin dashboard: switch columns & speed up
Columns switched: "domain" (renamed to "host") and "slug" because slug
is more important.

Speed up: counting number of instructors by annotation, not actual
`SELECT COUNT(*)...` queries for each event.
This actually produces twice as many instructors in cases of cancelled
workshops, but it's not a real issue, because we mainly care about this
number being zero. Still, a bug, so perhaps will be fixed one day.