Fix #1154 -- autocomplete fields no longer ignore invalid values
When you entered an invalid vaue into an autocomplete field (so that the field is red) and submit the form, invalid value is ignored and NULL was saved. No warning was displayed. This was confusing.

This commit prevents form submission if there is an invalid value in any autocomplete field. On top of that, the invalid field is focused on form submission.

There is no regression test, because we don't have end-to-end tests with javascript support (like Selenium).