Update merging objects behavior re. protected relations
To reflect Django's requirement for related objects to have explicit
`on_delete` parameter set in models, AMY's models were updated.

This however exposed us to an error in merging objects utility: the
"other" object (the one that has to be deleted at the end of merging)
cannot be deleted when some of its related objects weren't deleted.

This commit updates and extends the merging behavior to reflect this.