Add `publish_profile` to API filters (and discover some bugs!)
With the recently introduced `Person.publish_profile` field (to allow
users opt-in into sharing their profiles on The Carpentries website)
I discovered some bugs associated with the way `export-instructors`
queryset worked.
Namely, it was possible to associate a normal (non-instructor) user
with some airport to show this specific user in `export-instructors`
API endpoint!
A special Django query (using a `Prefetch` objects) was introduced to
accomodate this issue, but unfortunately there's a caveat: when a normal
user is associated with an airport, the airport will show empty list
of instructors (we'd rather like it to not show on the list at all).