Move jQuery import to <head>
This is required, because previously all JS was included on bottom of
the `<body>`. Now, however, with `form.media` gone from `extrajs` Django
template block, form media may be included somewhere in the middle of
`<body>` and trigger errors when `$()` is missing.

The solution for this issue is to move jQuery import before `<body>`.