Recalculate auto score upon M2M change in TrainingRequest
This required a bit of changes in the `workshops` application structure.
According to the Django documentation, signals should be connected in
AppConfig.ready() method, but since this project is quite old the
application config for `workshops` didn't exist. It was necessary to add
it.

The connected signals are:
* M2M change on `domains` field
* M2M change on `previous_involvement` field

The issue with this approach is that we only control one side of the M2M
relation - the TrainingRequest's side. If changes are to be made on
KnowledgeDomain's or Role's sides, then we'd have to make
N recalculations and saves. We therefore opt not to.