Display curricula on event's details page, and in EventForm
EventForm code was reordered to look more like other forms and less
frankenstein-y.

Additionally EventForm included a validation for curricula: if some
were selected, their name was matched to a tag name and an error
would be displayed if that tag was not assigned in the other form
field ('tags').