Add EventCreateForm inheriting from EventForm, send & receive signal
The new form, EventCreateFrom, is used in multiple places, e.g. when
accepting a workshop request. This form overrides `save` method to send
a signal when underlying model instance is being saved.

This is not the only change regarding EventForm. EventCreateForm
additionally contains a `comment` field used for storing comment
contents that will be saved as first comment.