Fiscal: add create forms for Organization and Membership
The new forms support comment field for adding comments. The contents
are send via a signal to the receiver, which adds a comment to that
object being saved in the form.

This solution enables us to add field `comment`, which is not bound to
a real model field, and use it for adding real comment once the object
begins existing.