Add comment field for EventForm as well
This is dictated by the fact, that we use 'Update from URL' functionality
which loads the workshop website and reads instructor and helper names
from it; then it loads these names into a comment (previously 'notes'
field).

Now we want to still have this information, so the decision is to bring
the 'comment' field back to EventForm and make it quite similar to
EventCreateForm.

Some things ended up simplified, like testing form comments or form
inheritance.

This fixes #1339.