Add fake Extcomments Django Comments app to overwrite default form
The Django Comments form is not possible to overwrite, unless this trick
is made; the application allows to specify a different comments
application as a sort-of extension of the default model and form.

We wanted only to change the form, so a minimalistic "extcomments" app
was created and Django Comments was set up to work with it.