Change place where jQuery is included in base.html
There are some issues on the main production server that can only be
resolved, for the countless time, by changing how jQuery is imported.

This time the change is to move main jQuery import from `<head>` to top
of `<body>`, and to move all other imports from bottom of the `<body>`
to the top.