Event.attendance -> manual_attendance, + attendance calculation
Attendance is now both a cached_property and an annotation.
If the annotation is not present, then we can access the
`attendance` @cached_property.

Unfortunately due to some weird Django bug (ticket unknown), it wasn't
possible to alter EventManager.get_query, so that all queries have
this annotation. Therefore we have to manually add
`Event.objects.attendance()`.

In case of single objects w/o this annotation, the `.attendance` is
a @cached_property and accessible.