DAL TagSelect2: fix error when corresponding tag missing from form data
When initiating `Form(dict())`, an error would occur: field using
TagSelect2 widget would try to get values provided, but under specified
field name, data was empty. This led to errors.

Additionally, this commit removes other errors linked to
`requested_workshop_types` widget in WorkshopRequestBaseForm. It was
supposed to be RadioSelect widget, but in the coding fever the old
widget override was left in the code.
The same widget shouldn't be marked as required by the form, but rather
an effect of `clean()` validation.