EventForm: build expected tags in clean_curricula differently
Previously we were using name of the curriculum (split by "-"). Now we
can use curriculum.carpentry value, which corresponds to Tag name,
unless it's mix&match -> then tag name is "Circuits". And this is what
the method now does.

A new test case was added to ensure this mechanism is working correctly
with Mix&Match curriculum.