Add Dockerfile and .dockerignore
The dockerfile creates AMY image but ultimately for development
purposes, as it only runs development server.
Installation is prepared for installing PIL.

TODO:
* install JS dependencies,
* install deployment services,
* run deployment services.

The `.dockerignore` file, used to copy data from local directory to
the container's directory, is based on `.gitignore`, with a single
exception: it doesn't ignore `git_version.py` file.