Tags: add priority field
The new field represents a sorting order. Tags with lower value will
appear first on the list (ascending order). If some tags have the same
sorting value, then they're sorted by name (A-Z).

By default, new Tags will have priority 0.

We're taking into account a new tag, Pilot, which is only present in
production server. With it in mind, the list of tags was visually
prolonged (14 -> 15) to prevent scrollbar.

A proposed order of tags is as follows:
  Circuits, Pilot, SWC, DC, LC, ITT, TTT, LMO, LSO, cancelled,
  unresponsive, stalled, WiSE, online, hackathon

Admins are able to easily change the order from Django Admin page.

This fixes #1574.