Non-transactional DB error when matching Tr. Req.
During matching Training Request with existing person in database,
this person's details are updated from the Request.

It may happen that there's a conflict in our DB. For example,
a duplicate person with the same email as in the Training Request.

In this case the rewriting of person's data will trigger
IntegrityError, and before this commit it wasn't in a transaction
and could result in 500 Server Error.