Additional changes in how SOS imports data to Event Form
Event Form is used when accepting Self-Organised Submissions
& creating new event at the same time.

This form is prefilled with some values from SOS object.

Previously, when an URL was provided, and "Import from URL" failed
to load tags from that URL, other values that don't depend on URL
failed to load as well. But now they always load.