[1754] Duration warning on membership details page
Show membership duration warning if the membership was created with
invalid duration (ie. != full year).

The year calculation is based on replacing year with start_year+1 and
comparing to selected end date.

The warning will only show up after creating the membership, not on
every request to the membership details page.