[#1939] Make affiliated organisations symmetrical
Symmetrical option makes affiliated organisation automatically
"point back" to the organisation that initiated relationship.

For example, previously we could relate organistations: A->B.
To relate B with A we'd have to add another relation entry B->A.

After the change, once we set A->B then B->A is automatically set
(we end up with symmetrical relation `A<->B`).