[#1930] Unique constraint in membership formsets
The solution is not perfect.

In a nutshell, it turned out that complete unique validation only
happens when user provides membership field in the formset's forms.
All attempts to provide this value on the backend side failed in one way
or another. The closest I got was validation error when user provided
values that were already on file, but then if they provided 2 the same
values that weren't on the file, the validation would pass - but
database would throw integrity error.