[#2540] Fix bug by removing a feature
Bug was first reported by @elichad in https://github.com/carpentries/amy/pull/2551#pullrequestreview-1737378977

The bug was that page didn't render because of some weird
Python inspection stuff. As it turned out, that inspection was happening
in Django template system when trying to display base class for Python
date.

It was not possible to get rid of that error in a sane matter, so
I decided to remove the feature linking models with their documentation.